Move storage from media to lib. Added interface and configurable filesystem media path.
